Recover the Ocean Force Point electric-floor puzzle from the port
The port played this puzzle and found the whole family was named after the wrong
level and the wrong mechanic. What the tree called a DragonRock "spell puzzle"
with a "mode table" and a per-mode "required score" is Ocean Force Point's
electric floor: a grid of tiles where one tile per row is safe, the pad shows
the solution while pressed, and stepping wrong zaps you.

So gDFPLevelControlPuzzleValues is gDFPLevelControlSafeFloorTiles,
gDfpfloorbarModeTable is gDFPFloorbarSafeFloorTiles, DfpLevelControlInterface's
copyPuzzleValues is copySafeFloorTiles, and modeIndex/requiredScore across
chuka, dfpfloorbar and dfpwallbar are rowIndex/safeTileIndex. ChukaState's
linkedObject is the levelController GameObject* it always held, dfpfloorbar's
completionGameBit/active are loweredGameBit/lowered, and its lastSequenceValue
is previousShowSolutionState. DfpLevelControlState's timer is zappedTimer, its
sfxLatch and flags07 are one previousSfxState bitfield, and gameBitLatches[4] is
the s32 musicLatchMask the three GameBitLatch_Update calls share.

Eight gamebits are named, and one corrected: 0x5E4 was
GAMEBIT_DRBOT_SpellPuzzleActive and is GAMEBIT_OFP_PuzzlePadShowSolution. New
ids go in the unordered section in ascending order per the house rule.

Four function renames, all one-unit blast radius, done through symbols.txt with
pairing_check clean and the FORCEACTIVE warning count unmoved at zero:
DFP_LevelControl_updateMapAct1/2 -> updateAct1/2, sequenceCallback ->
animCallback, copyPuzzleValues -> copySafeFloorTiles.

Also from this batch: LightfootButtonTimingControlState (437) is a padded view of
LightfootControlState and is now expressed as one, with moveIds, moveSpeeds and
its four timers named; DfpTargetBlockCollisionPoints moves out of door.h into
DfpTargetBlockState as floorPoints[8]/floorPointCount, with the walk's offsets
now offsetof-derived instead of magic constants; DFP_RotateP's effect handles,
344's barrel generators, DBstealerwo's object cursors and track_dolphin's
objGetAllOfType result are GameObject** rather than int*/u32*; DBstealerwo's
`int b8` scratch becomes a GroundBaddieState*/DbStealerwormControl* pair; and
DLL 0x299's Dll299Vtable is the real DllA6Interface.

main.dol's sha1 (e750e8e8) is unchanged and no unit's fuzzy moved: 99.817474,
matched_code 2524296, matched_data 1204361, complete_units 916,
matched_functions 9291. The six objects whose md5 shifts all pass
unit_bytes_check with 0 differ.

Six shapes the port simplified had to be put back, each found from the target
asm: DFP_LevelControl_copySafeFloorTiles unrolls by three with (s16) index
casts; animCallback reads its timer through a local; DFP_RotateP's per-handle
position update is a macro retail inlines, not a function, and its
ensureEffectHandlePair keeps two separate cursor computations; DFP_Lightni_init
and dfppowersl_init write their clamped radius and object ids back into the
placement rather than into locals. 572_DFP_PowerSl is left alone entirely - its
diff is an early-return restyle that costs two functions and recovers nothing.
The 626 unsequenced-assignment fix was re-measured and still costs 100 ->
99.826, as CLAUDE.md records.

+++ src/dlls/engine/15/15.c
@@ -459,7 +459,7 @@ void player_rotateTowardEnemy(GameObject* obj, int* ctx, int spd)
     enemy = (GameObject*)((BaddieState*)ctx)->targetObj;
     if (enemy != 0)
     {
-        if ((u32)enemy->anim.parent == (u32)obj->anim.parent)
+        if (enemy->anim.parent == obj->anim.parent)
         {
             dx = enemy->anim.localPosX - obj->anim.localPosX;
             dz = enemy->anim.localPosZ - obj->anim.localPosZ;
